tweak: move less-common code to lib folder
