+++ src/compiler/compile/expression_float.rs
@@ -1,7 +1,7 @@
 use crate::compiler::opcode::OpCode;
 use crate::compiler::Compiler;
-use crate::object::float;
-use crate::object::Object;
+use crate::lib::object::float;
+use crate::lib::object::Object;
 use crate::parser::expression::float::Float;
 
 pub fn compile_expression_float(compiler: &mut Compiler, flt: Float) -> Option<String> {
