+++ src/compiler/compile/expression_integer.rs
@@ -1,7 +1,7 @@
 use crate::compiler::opcode::OpCode;
 use crate::compiler::Compiler;
-use crate::object::integer;
-use crate::object::Object;
+use crate::lib::object::integer;
+use crate::lib::object::Object;
 use crate::parser::expression::integer::Integer;
 
 pub fn compile_expression_integer(compiler: &mut Compiler, int: Integer) -> Option<String> {
