+++ src/lib/README.md
@@ -1,0 +1,2 @@
+# Library
+This folder contains modules which are used by the Loop runtime
