+++ src/lib/object/integer.rs
@@ -1,5 +1,5 @@
-use crate::object::float::Float;
-use crate::object::ObjectTrait;
+use crate::lib::object::float::Float;
+use crate::lib::object::ObjectTrait;
 
 #[derive(Copy, Clone, Debug, PartialEq)]
 pub struct Integer {
