+++ src/lib/object/null.rs
@@ -1,4 +1,4 @@
-use crate::object::ObjectTrait;
+use crate::lib::object::ObjectTrait;
 
 #[derive(Copy, Clone, Debug, PartialEq)]
 pub struct Null {}
