+++ src/vm/mod.rs
@@ -7,9 +7,9 @@ use crate::compiler::definition::lookup_op;
 use crate::compiler::instructions::{read_uint32, read_uint8};
 use crate::compiler::opcode::OpCode;
 use crate::compiler::Bytecode;
-use crate::object::function::{CompiledFunction, Function};
-use crate::object::null::Null;
-use crate::object::Object;
+use crate::lib::object::function::{CompiledFunction, Function};
+use crate::lib::object::null::Null;
+use crate::lib::object::Object;
 use crate::vm::frame::{build_frame, Frame};
 use crate::vm::function::{run_function, run_function_stack};
 use crate::vm::suffix::run_suffix_expression;
